requirements.txt
streamlit
# torch
# transformers
# peft
tensorflow
gensim
pyspellchecker
nltk
json
numpy
# huggingface_hub
